add integration tests for upload-unbound-stream
