+++ test/integration/test_autocomplete.py
@@ -8,7 +8,6 @@
 #
 ######################################################################
 
-import os
 import sys
 
 import pexpect
@@ -29,49 +28,10 @@ echo "Just testing if we don't replace existing script" > /dev/null
 
 
 @pytest.fixture(scope="session")
-def homedir(tmp_path_factory):
-    yield tmp_path_factory.mktemp("test_homedir")
-
-
-@pytest.fixture(scope="session")
 def bashrc(homedir):
     bashrc_path = (homedir / '.bashrc')
     bashrc_path.write_text(BASHRC_CONTENT)
     yield bashrc_path
-
-
-@pytest.fixture(scope="session")
-def b2_in_path(tmp_path_factory):
-    """
-    Create a dummy b2 executable in a temporary directory and add it to PATH.
-
-    This allows us to test the b2 command line tool even if tested `b2` package was not installed.
-    """
-
-    tempdir = tmp_path_factory.mktemp("temp_bin")
-    temp_executable = tempdir / "b2"
-    with open(temp_executable, "w") as f:
-        f.write(
-            f"#!{sys.executable}\n"
-            "import sys\n"
-            f"sys.path.insert(0, {os.getcwd()!r})\n"  # ensure relative imports work even if command is run in different directory
-            "from b2.console_tool import main\n"
-            "main()\n"
-        )
-
-    temp_executable.chmod(0o755)
-
-    original_path = os.environ["PATH"]
-    new_path = f"{tempdir}:{original_path}"
-    yield new_path
-
-
-@pytest.fixture(scope="module")
-def env(b2_in_path, homedir, monkey_patch):
-    monkey_patch.setenv('PATH', b2_in_path)
-    monkey_patch.setenv('HOME', str(homedir))
-    monkey_patch.setenv('SHELL', "/bin/bash")  # fix for running under github actions
-    yield os.environ
 
 
 @pytest.fixture(scope="module")
